HTML and CSS for Header and First Button
Updated the HTML to have the proper layout and links for the first state of the website. The names have links but only mine works right now. I have some input fields but there is still more work to do for the appealing time.
Spent a good amount of time for the styling of the button and header, will need to work on changing the percentages for the sizes of some blocks.
There is some basic colors and font for the whole website.
The rest of the page still needs proper styling and sectioned off.
Added a pear image for the icon of the tab, (also new tab name).
Lastly, added a markdown file that contains some resources I used.