Upload Section Styled, Small HTML Tweaks, etc
The big orange button grew a bit. I removed the margins from the body of the page to make the header fit better and now it sticks to the page.
Slight padding to the bottom of the authors and different radius only in bottom corners,
The sections for uploading a picture got styled with a new upload icon.
I moved the upload div into the same div as the big button for scripting later. I also added a meta tag in the head and a submit button next to the URL.
Flex was added to the URL div so that way there was a url input and submit button.