Footer Section and Preview Image
I primarily styled the footer for most of the time, there was no real extra work in HTML for the footer other than the inclusion of the arrow down image.
About Section button got new style with the icon, the paragraphs got headings and extra sentences, flex added to the div so the content spaces out properly.

New div created that will house the image preview of selected image and contains the buttons from the design. This changed the HTML page the most, but the styling was very simple as there was already the previous class created that was applied to this new div. The styling of the buttons was created as one class and the background colors were treated separately with IDs.