Drag and Drop + About Section Functionality
JavaScript updated to support for dragging and dropping image files (only png and jpg) into the upload button label on top of also prompting a file select window.
I separated some repetitive code to make the code look more clean and readable. I created a function that would check the length of the list of files being inputted as to only allow one file. There is also a function that does the procedure for saving the image as a URL and moving to the next step of the process to preview the image.

Move the script tag in the HTML to below the Footer to allow for the script to be able to access footer elements. Now the section with all the "About This Project" text appears when clicked and also disappears when clicked again. For looks I will rotate the arrow whenever the section is open or closed.