URL Submission (W.I.P.) + More Resources
Added a few event handlers. One if for the URL Submit button so that a URL will create an image for the next step in the process. Currently it only works on actual image URLs, and it does not create files as of yet.
There is also another event handler for after submitting an image and moving onto the next step. All it does is create an alert and turn off the current div. Will work on creating the next section to display an image and some text plus a button to return to the first step.
Some resources got added that were big influences for creating some Image file error catching (to be implemented later) in the meantime before fetching actual image files.